requirements.txt
python-dotenv
tensorflow # Atau pytorch, tergantung pilihanmu
requests
